# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cinema Booking System API",
        "description": "API documentation for Cinema Booking System",
        "contact": {
            "email": "support@cinemat.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "{protocol}://{host}",
            "description": "API Server",
            "variables": {
                "protocol": {
                    "enum": [
                        "http",
                        "https"
                    ],
                    "default": "http"
                },
                "host": {
                    "default": "localhost:8089"
                }
            }
        }
    ],
    "paths": {
        "/api/movies/now-playing": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy danh sách phim đang chiếu",
                "operationId": "7b824264c7197e682451673a98572178",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Số trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "description": "Ngôn ngữ",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lấy danh sách phim đang chiếu thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/upcoming": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy danh sách phim sắp chiếu",
                "operationId": "ac3aec442fd49b864f67073f7af543c4",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/popular": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy danh sách phim phổ biến",
                "operationId": "e77d08ecceeb5d37263fb4884541d639",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/{movieId}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy chi tiết phim",
                "operationId": "7896eaf5c7ada516b96a4456be372fb0",
                "parameters": [
                    {
                        "name": "movieId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/search": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Tìm kiếm phim",
                "operationId": "ef2b563a3ef02baa505461bbcac1e800",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "400": {
                        "description": "Thiếu từ khóa tìm kiếm"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/genres": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy danh sách thể loại phim",
                "operationId": "b20597cba9bc0d02a049a8c12d63d91b",
                "parameters": [
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
        "/api/movies/genre/{genreId}": {
            "get": {
                "tags": [
                    "Movies"
                ],
                "summary": "Lấy danh sách phim theo thể loại",
                "operationId": "e088dc5fef7fd9fab51c08b196727c58",
                "parameters": [
                    {
                        "name": "genreId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "language",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "vi-VN"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "500": {
                        "description": "Lỗi server"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}